Time: 26 ms (81.52%) | Memory: 107.5 MB (30.95%) - LeetSync

+++ 2245-destroying-asteroids/destroying-asteroids.java
@@ -2,8 +2,8 @@ class Solution {
     public boolean asteroidsDestroyed(int mass, int[] asteroids) {
         Arrays.sort(asteroids);
         
-        long bigmass=(long)mass;
-        
+        long bigmass=mass;
+
         for(int i=0;i<asteroids.length ;i++){
 
             if(asteroids[i]<=bigmass) bigmass+=asteroids[i];
